Remove Material element class
This changelist removes the legacy Material element class, and merges MaterialNode.cpp/h into Material.cpp/h and Look.cpp/h.
