+++ source/MaterialXCore/Document.h
@@ -100,54 +100,6 @@ class Document : public GraphElement
     vector<PortElementPtr> getMatchingPorts(const string& nodeName) const;
 
     /// @}
-    /// @name Material Elements
-    /// @{
-
-    /// Add a Material to the document.
-    /// @param name The name of the new Material.
-    ///     If no name is specified, then a unique name will automatically be
-    ///     generated.
-    /// @return A shared pointer to the new Material.
-    MaterialPtr addMaterial(const string& name = EMPTY_STRING)
-    {
-        return addChild<Material>(name);
-    }
-
-    /// Return the Material, if any, with the given name.
-    MaterialPtr getMaterial(const string& name) const
-    {
-        return getChildOfType<Material>(name);
-    }
-
-    /// Return a vector of all Material elements in the document.
-    vector<MaterialPtr> getMaterials() const
-    {
-        return getChildrenOfType<Material>();
-    }
-
-    /// Remove the Material, if any, with the given name.
-    void removeMaterial(const string& name)
-    {
-        removeChildOfType<Material>(name);
-    }
-
-    /// @}
-    /// @name Material Node Utilities
-    /// @{
-
-    /// Return the Material node, if any, with the given name.
-    NodePtr getMaterialNode(const string& name) const
-    {
-        return getNode(name);
-    }
-
-    /// Return a vector of all Material nodes in the document.
-    vector<NodePtr> getMaterialNodes() const
-    {
-        return getNodesOfType(MATERIAL_TYPE_STRING);
-    }
-
-    /// @}
     /// @name GeomInfo Elements
     /// @{
