+++ source/MaterialXTest/MaterialXCore/Look.cpp
@@ -13,28 +13,14 @@ TEST_CASE("Look", "[look]")
 {
     mx::DocumentPtr doc = mx::createDocument();
 
-    // Create a material and look.
-    mx::MaterialPtr material = doc->addMaterial();
+    // Create a look.
     mx::LookPtr look = doc->addLook();
-    REQUIRE(doc->getMaterials().size() == 1);
     REQUIRE(doc->getLooks().size() == 1);
 
-    // Bind the material to a geometry string.
-    mx::MaterialAssignPtr matAssign1 = look->addMaterialAssign("matAssign1", material->getName());
-    matAssign1->setGeom("/robot1");
-    REQUIRE(matAssign1->getReferencedMaterial() == material);
-    REQUIRE(material->getGeometryBindings("/robot1").size() == 1);
-    REQUIRE(material->getGeometryBindings("/robot2").size() == 0);
-
-    // Bind the material to a geometric collection.
-    mx::MaterialAssignPtr matAssign2 = look->addMaterialAssign("matAssign2", material->getName());
+    // Create a geometric collection.
     mx::CollectionPtr collection = doc->addCollection();
     collection->setIncludeGeom("/robot2");
     collection->setExcludeGeom("/robot2/left_arm");
-    matAssign2->setCollection(collection);
-    REQUIRE(material->getGeometryBindings("/robot2").size() == 1);
-    REQUIRE(material->getGeometryBindings("/robot2/right_arm").size() == 1);
-    REQUIRE(material->getGeometryBindings("/robot2/left_arm").size() == 0);
 
     // Create a property assignment.
     mx::PropertyAssignPtr propertyAssign = look->addPropertyAssign();
@@ -76,7 +62,6 @@ TEST_CASE("Look", "[look]")
     // Create an inherited look.
     mx::LookPtr look2 = doc->addLook();
     look2->setInheritsFrom(look);
-    REQUIRE(look2->getActiveMaterialAssigns().size() == 2);
     REQUIRE(look2->getActivePropertySetAssigns().size() == 1);
     REQUIRE(look2->getActiveVisibilities().size() == 1);
 
@@ -88,7 +73,6 @@ TEST_CASE("Look", "[look]")
 
     // Disconnect the inherited look.
     look2->setInheritsFrom(nullptr);
-    REQUIRE(look2->getActiveMaterialAssigns().empty());
     REQUIRE(look2->getActivePropertySetAssigns().empty());
     REQUIRE(look2->getActiveVisibilities().empty());
 }
