+++ source/PyMaterialX/PyMaterialXCore/PyDocument.cpp
@@ -25,13 +25,6 @@ void bindPyDocument(py::module& mod)
         .def("getNodeGraphs", &mx::Document::getNodeGraphs)
         .def("removeNodeGraph", &mx::Document::removeNodeGraph)
         .def("getMatchingPorts", &mx::Document::getMatchingPorts)
-        .def("addMaterial", &mx::Document::addMaterial,
-            py::arg("name") = mx::EMPTY_STRING)
-        .def("getMaterial", &mx::Document::getMaterial)
-        .def("getMaterials", &mx::Document::getMaterials)
-        .def("removeMaterial", &mx::Document::removeMaterial)
-        .def("getMaterialNode", &mx::Document::getMaterialNode)
-        .def("getMaterialNodes", &mx::Document::getMaterialNodes)
         .def("addGeomInfo", &mx::Document::addGeomInfo,
             py::arg("name") = mx::EMPTY_STRING, py::arg("geom") = mx::UNIVERSAL_GEOM_NAME)
         .def("getGeomInfo", &mx::Document::getGeomInfo)
