+++ source/PyMaterialX/PyMaterialXCore/PyElement.cpp
@@ -86,13 +86,12 @@ void bindPyElement(py::module& mod)
         .def("getRoot", static_cast<mx::ElementPtr(mx::Element::*)()>(&mx::Element::getRoot))
         .def("getDocument", static_cast<mx::DocumentPtr(mx::Element::*)()>(&mx::Element::getDocument))
         .def("traverseTree", &mx::Element::traverseTree)
-        .def("traverseGraph", &mx::Element::traverseGraph,
-            py::arg("material") = nullptr)
+        .def("traverseGraph", &mx::Element::traverseGraph)
         .def("getUpstreamEdge", &mx::Element::getUpstreamEdge,
-            py::arg("material") = nullptr, py::arg("index") = 0)
+            py::arg("index") = 0)
         .def("getUpstreamEdgeCount", &mx::Element::getUpstreamEdgeCount)
         .def("getUpstreamElement", &mx::Element::getUpstreamElement,
-            py::arg("material") = nullptr, py::arg("index") = 0)
+            py::arg("index") = 0)
         .def("traverseInheritance", &mx::Element::traverseInheritance)
         .def("setSourceUri", &mx::Element::setSourceUri)
         .def("hasSourceUri", &mx::Element::hasSourceUri)
@@ -117,7 +116,6 @@ void bindPyElement(py::module& mod)
         BIND_ELEMENT_FUNC_INSTANCE(GeomProp)
         BIND_ELEMENT_FUNC_INSTANCE(Implementation)
         BIND_ELEMENT_FUNC_INSTANCE(Look)
-        BIND_ELEMENT_FUNC_INSTANCE(Material)
         BIND_ELEMENT_FUNC_INSTANCE(MaterialAssign)
         BIND_ELEMENT_FUNC_INSTANCE(Node)
         BIND_ELEMENT_FUNC_INSTANCE(NodeDef)
